+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Bug fixes
 * The Deku Shield pot in the Spirit Temple is no longer shuffled when both `Fix Broken Drops` and `Include Empty Pots` are off.
+* For consistency with other location types, excluded pot/crate/beehive/wonderitem/etc. locations no longer behave as if their shuffle setting was off.
 
 ## Other changes
 * Big poe souls can now be collected while riding Epona.